+++ movies.cpp
@@ -1,7 +1,7 @@
 #include "movies.h"
 
 
-void print_alphabetical(map<string, int> mbin)
+void print_alphabetical(map<string, double> mbin)
 {
     for(auto movie : mbin) //since std::map is automatically sorted in ascending order this should be alphabetical order
     {
